+++ R/extract_covariates.R
@@ -4,7 +4,8 @@
 #' steps.
 #' @template track_xy_star_steps
 #' @param covariates `[RasterLayer,RasterStack,RasterBrick]` \cr The
-#'   (environmental) covariates.
+#'   (environmental) covariates. For `extract_covariates_var_time` the argument
+#'   `covariates` need to have a `z`-columne (i.e. the time stamp).
 #' @param where `[character(1)="end"]{"start", "end", "both"}` \cr For `steps` this
 #'   determines if the covariate values should be extracted at the beginning or
 #'   the end of a step. or `end`.
@@ -103,3 +104,168 @@ extract_covariates_along.steps_xy <- function(x, covariates, ...) {
     stop("covariates: need to be a Raster*.")
   }
 }
+
+
+# Extract covariates varying time -----------------------------------------
+#' @rdname extract_covariates
+#' @param when `[character(1)="any"]{"any", "before", "after"}` \cr Specifies for
+#'  for `extract_covariates_var_time` whether to look before, after or in both
+#'  direction (`any`) for the temporally closest environmental raster.
+#' @param max_diff `[Period(1)]` \cr The maximum time difference between a relocation
+#'  and the corresponding raster. If no rasters are within the specified max.
+#'  distance `NA` is returned.
+#' @param name_covar `[character(1)="time_var_covar"]` \cr The name of the new column.
+#' @export
+#' @examples
+#' # Simulate some dummy data
+#' # Hourly data for 10 days: 24 * 10
+#' set.seed(123)
+#' path <- data.frame(x = cumsum(rnorm(240)),
+#'               y = cumsum(rnorm(240)),
+#'               t = lubridate::ymd("2018-01-01") + hours(0:239))
+#'
+#' # dummy env data
+#' rs <- raster::raster(xmn = -50, xmx = 50, ymn = -50, ymx = 50, res = 1)
+#'
+#' # create dummy covars for each day
+#' rs <- raster::stack(lapply(1:10, function(i)
+#'   raster::setValues(rs, runif(1e4, i - 1, i))))
+#'
+#' # Env covariates are always taken at noon
+#' rs <- raster::setZ(rs, lubridate::ymd_hm("2018-01-01 12:00") + days(0:9))
+#'
+#' # Allow up to 2 hours after
+#' trk %>% extract_covariates_var_time(rs, max_time = hours(2), when = "after") %>%
+#'   print(n = 25)
+#' trk %>% extract_covariates_var_time(rs, max_time = hours(2), when = "before") %>%
+#'   print(n = 25)
+#' trk %>% extract_covariates_var_time(rs, max_time = hours(2), when = "any") %>%
+#'   print(n = 25)
+#'
+#' # We can use different time scales
+#' trk %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(2), when = "any", name_covar = "env_2h") %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(4), when = "any", name_covar = "env_4h") %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(6), when = "any", name_covar = "env_6h") %>%
+#'   print(n = 25)
+#'
+#' # We can use different time scales: after
+#' trk %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(2), when = "after", name_covar = "env_2h") %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(4), when = "after", name_covar = "env_4h") %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(6), when = "after", name_covar = "env_6h") %>%
+#'   print(n = 25)
+#'
+#' # We can use different time scales: before
+#' trk %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(2), when = "before", name_covar = "env_2h") %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(4), when = "before", name_covar = "env_4h") %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(6), when = "before", name_covar = "env_6h") %>%
+#'   print(n = 25)
+#'
+#' # The same works also for steps
+#' trk %>%
+#'   steps() %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(2), when = "before", name_covar = "env_2h") %>%
+#'   print(n = 25)
+#'
+#' # also with start and end
+#' trk %>%
+#'   steps() %>%
+#'   extract_covariates_var_time(
+#'     rs, max_time = hours(2), when = "before", name_covar = "env_2h",
+#'     where = "both") %>%
+#'   print(n = 25)
+#'
+extract_covariates_var_time <- function(x, ...) {
+  UseMethod("extract_covariates_var_time", x)
+}
+
+#' @export
+#' @rdname extract_covariates
+extract_covariates_var_time.track_xyt <- function(
+  x, covariates, when = "any", max_time,
+  name_covar = "time_var_covar", ...) {
+  x[name_covar] <- extract_covar_var_time_base(
+    cbind(x$x_, x$y_),
+    x$t_, covariates, when, max_time)
+  x
+}
+
+#' @export
+#' @rdname extract_covariates
+extract_covariates_var_time.steps_xyt <- function(
+  x, covariates, when = "any", max_time, name_covar = "time_var_covar",
+  where = "end", ...) {
+
+  if (where == "start") {
+    x[name_covar] <- extract_covar_var_time_base(
+      cbind(x$x1_, x$y1_),
+      x$t1_, covariates, when, max_time)
+  } else if (where == "end") {
+    x[name_covar] <- extract_covar_var_time_base(
+      cbind(x$x2_, x$y2_),
+      x$t2_, covariates, when, max_time)
+  } else if (where == "both") {
+    x[paste0(name_covar, "_start")] <- extract_covar_var_time_base(
+      cbind(x$x1_, x$y1_),
+      x$t1_, covariates, when, max_time)
+    x[paste0(name_covar, "_end")] <- extract_covar_var_time_base(
+      cbind(x$x2_, x$y2_),
+      x$t2_, covariates, when, max_time)
+  }
+  x
+}
+
+
+extract_covar_var_time_base <- function(
+  xy, t, covariates, when = "any",
+  max_diff) {
+
+  if (is.null(raster::getZ(covariates))) {
+    stop("Covariates do not have a Z column.")
+  }
+
+  if (!is(max_diff, "Period")) {
+    stop("`max_diff` is not of class `Period`.")
+  }
+  max_diff <- lubridate::period_to_seconds(max_diff)
+  t_covar <- as.numeric(as.POSIXct(raster::getZ(covariates)))
+  t_obs <- as.numeric(t)
+
+  # Fun to find closest point
+  which_rast <- function(t_diffs, where, max_diff) {
+    wr <- if (when == "after") {
+      which.min(t_diffs[t_diffs >= 0])
+    } else if (when == "before") {
+      which.min(abs(t_diffs[t_diffs <= 0])) + sum(t_diffs > 0)
+    } else if (when == "any") {
+      which.min(abs(t_diffs))
+    }
+    if (length(wr) == 0) {
+      NA
+    } else if (max_diff < abs(t_diffs[wr])) {
+      NA
+    } else {
+      wr
+    }
+  }
+
+  wr <- sapply(t_obs, function(x) which_rast(x - t_covar, when, max_diff))
+  ev <- raster::extract(covariates, cbind(xy))
+  cov_val <- ev[cbind(seq_along(wr), wr)]
+  return(cov_val)
+}
+
+
+