+++ man-roxygen/track_xyt.R
@@ -1,0 +1,1 @@
+#' @param x \code{[track_xyt]} \cr A track created with \code{make_track} that includes time.