+++ R/amt.R
@@ -1,5 +1,6 @@
 #' @importFrom dplyr arrange filter group_by select summarize summarise ungroup
 #' @importFrom dplyr distinct ungroup mutate mutate_at pull %>% bind_rows select_vars n
+#' @importFrom dplyr count
 #' @importFrom graphics plot legend
 #' @importFrom grDevices adjustcolor
 #' @importFrom lazyeval lazy_dots lazy
@@ -34,6 +35,8 @@ utils::globalVariables(c("burst_", "step_id_", "t_", ".data", "ts")) # to omit C
 #' @export
 dplyr::arrange
 #' @export
+dplyr::count
+#' @export
 dplyr::filter
 #' @export
 dplyr::group_by
@@ -45,60 +48,41 @@ dplyr::summarise
 dplyr::summarize
 #' @export
 dplyr::ungroup
-
 #' @export
 dplyr::distinct
-
 #' @export
 dplyr::ungroup
-
 #' @export
 dplyr::mutate
-
 #' @export
 dplyr::pull
-
 #' @export
 sp::CRS
-
 #' @export
 lubridate::hours
-
 #' @export
 lubridate::minutes
-
 #' @export
 lubridate::seconds
-
 #' @export
 lubridate::days
-
 #' @export
 lubridate::weeks
-
 #' @export
 purrr::map
-
 #' @export
 survival::clogit
-
 #' @export
 survival::strata
-
 #' @export
 tidyr::nest
-
 #' @export
 tidyr::unnest
-
 #' @export
 tibble::tibble
-
 #' @export
 magrittr::`%>%`
-
 #' @export
 survival::Surv
-
 #' @export
 raster::raster