+++ NEWS.md
@@ -1,8 +1,11 @@
 # amt 0.0.6.0
 
 ## major changes
+- random_points with kde home ranges now work with different home range levels. 
 - random_steps now uses `units` to be explicit about the turning angle.
 - `bbox` is also available for steps and can return objects of `sf`.
+
+
 
 # amt 0.0.5.0
 ## major changes